@@ -1,7 +1,7 @@
 package org.codehaus.mojo.jaxb2.schemageneration.postprocessing.javadoc.wrappers;
 
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlElementWrapper;
+import jakarta.xml.bind.annotation.XmlElement;
+import jakarta.xml.bind.annotation.XmlElementWrapper;
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
@@ -11,13 +11,14 @@ import java.util.TreeSet;
 /**
  * @author <a href="mailto:lj@jguru.se">Lennart J&ouml;relid</a>, jGuru Europe AB
  */
-public class ExampleXmlWapperUsingJavaBeanAccessors implements Serializable {
+public class ExampleXmlWapperUsingJavaBeanAccessors implements Serializable
+{
 
     private List<String> strings;
     private SortedSet<Integer> integerSet;
 
-
-    public ExampleXmlWapperUsingJavaBeanAccessors() {
+    public ExampleXmlWapperUsingJavaBeanAccessors()
+    {
 
         this.strings = new ArrayList<String>();
         this.integerSet = new TreeSet<Integer>();
@@ -26,9 +27,10 @@ public class ExampleXmlWapperUsingJavaBeanAccessors implements Serializable {
     /**
      * JavaBean getter for a List containing some strings.
      */
-    @XmlElementWrapper(name = "foobar")
-    @XmlElement(name = "aString")
-    public List<String> getStrings() {
+    @XmlElementWrapper( name = "foobar" )
+    @XmlElement( name = "aString" )
+    public List<String> getStrings()
+    {
         return strings;
     }
 
@@ -36,8 +38,9 @@ public class ExampleXmlWapperUsingJavaBeanAccessors implements Serializable {
      * JavaBean getter for a SortedSet containing Integers.
      */
     @XmlElementWrapper
-    @XmlElement(name = "anInteger")
-    public SortedSet<Integer> getIntegerSet() {
+    @XmlElement( name = "anInteger" )
+    public SortedSet<Integer> getIntegerSet()
+    {
         return integerSet;
     }
 }
